+++ src/components/support-form/SupportForm.tsx
@@ -3,9 +3,10 @@ import { useTranslation } from 'react-i18next'
 import { makeStyles, Theme, createStyles, withStyles } from '@material-ui/core/styles'
 import { Container, Stepper, Step, StepLabel, StepConnector, Button } from '@material-ui/core'
 import { FormikErrors } from 'formik'
-import * as yup from 'yup'
 
 import useForm from 'common/form/useForm'
+import { validationSchema } from './helpers/validation-schema'
+import { Steps, Step as StepType } from './helpers/support-form.models'
 import Layout from 'components/layout/Layout'
 import StepIcon from './StepperIcon'
 import GeneralInfo from './steps/GeneralInfo'
@@ -14,7 +15,7 @@ import ThankYou from './steps/ThankYou'
 import GDPR from './steps/GDRP'
 import Roles from './steps/Roles'
 
-import { SupportFormData } from './types/SuportFormData'
+import { SupportFormData } from './helpers/support-form.models'
 
 const ColorlibConnector = withStyles({
   alternativeLabel: {
@@ -56,121 +57,6 @@ const useStyles = makeStyles((theme: Theme) =>
   }),
 )
 
-type Step = {
-  label: string
-  component: JSX.Element
-}
-
-export enum Steps {
-  NONE = -1,
-  ROLES = 0,
-  QUESTIONS = 1,
-  INFO = 2,
-  GDPR = 3,
-}
-function checkboxChecked(this: any, value: { [key: string]: boolean }) {
-  const { path, createError } = this
-  if (!Object.values(value).filter((value) => value).length) {
-    return createError({ path, message: 'Must have at least one checked box' })
-  }
-  return true
-}
-
-const validationSchema: yup.SchemaOf<SupportFormData> = yup.object().shape({
-  terms: yup
-    .bool()
-    .required()
-    .test('checkboxChecked', 'Please check one checkbox', (value) => {
-      return value
-        ? true
-        : new yup.ValidationError('Please check one checkbox', null, 'checkboxChecked')
-    }),
-  info: yup
-    .object()
-    .shape({
-      email: yup.string().email().required(),
-      name: yup.string().required(),
-      phone: yup.string().required(),
-      address: yup.string().required(),
-    })
-    .required(),
-  roles: yup
-    .object()
-    .shape({
-      benefactor: yup.bool(),
-      partner: yup.bool(),
-      volunteer: yup.bool(),
-      associationMember: yup.bool(),
-      promoter: yup.bool(),
-    })
-    .required()
-    .test('checkboxChecked', 'Must have at least one checked box', checkboxChecked),
-  benefactor: yup.object().when('roles.benefactor', {
-    is: true,
-    then: yup
-      .object()
-      .shape({
-        campaignBenefactor: yup.bool(),
-        platformBenefactor: yup.bool(),
-      })
-      .test('checkboxChecked', 'Must have at least one checked box', checkboxChecked),
-  }),
-  associationMember: yup.object().when('roles.associationMember', {
-    is: true,
-    then: yup
-      .object()
-      .shape({
-        isMember: yup.bool(),
-      })
-      .test('checkboxChecked', 'Must have at least one checked box', checkboxChecked),
-  }),
-  partner: yup.object().when('roles.partner', {
-    is: true,
-    then: yup
-      .object()
-      .shape({
-        npo: yup.bool(),
-        bussiness: yup.bool(),
-        other: yup.bool(),
-        otherText: yup.string().when('partner.other', {
-          is: true,
-          then: yup.string(),
-        }),
-      })
-      .test('checkboxChecked', 'Must have at least one checked box', checkboxChecked)
-      .test('CustomValidation', 'Custom validation', function (this: any, values: any) {
-        const { path, createError } = this
-        return values.other && !values.otherText
-          ? createError({ path, message: 'field is required' })
-          : true
-      }),
-  }),
-  volunteer: yup.object().when('roles.volunteer', {
-    is: true,
-    then: yup.object().shape({
-      areas: yup.array().of(yup.string()),
-    }),
-  }),
-  promoter: yup.object().when('roles.promoter', {
-    is: true,
-    then: yup
-      .object()
-      .shape({
-        mediaPartner: yup.bool(),
-        ambassador: yup.bool(),
-        other: yup.bool(),
-        otherText: yup.string(),
-      })
-      .test('checkboxChecked', 'Must have at least one checked box', checkboxChecked)
-      .test('CustomValidation', 'Custom validation', function (this: any, values: any) {
-        const { path, createError } = this
-        return values.other && !values.otherText
-          ? createError({ path, message: 'field is required' })
-          : true
-      }),
-  }),
-})
-
 const defaults: SupportFormData = {
   terms: false,
   info: {
@@ -210,7 +96,7 @@ const defaults: SupportFormData = {
   },
 }
 
-export default function Steppers(this: any) {
+export default function SupportForm(this: any) {
   const { t } = useTranslation()
   const classes = useStyles()
   const [activeStep, setActiveStep] = useState(0)
@@ -280,11 +166,11 @@ export default function Steppers(this: any) {
     return step === failedStep
   }
 
-  const isLastStep = (activeStep: number, steps: Step[]): boolean => {
+  const isLastStep = (activeStep: number, steps: StepType[]): boolean => {
     return activeStep === steps.length - 2
   }
 
-  const isThankYouStep = (activeStep: number, steps: Step[]): boolean => {
+  const isThankYouStep = (activeStep: number, steps: StepType[]): boolean => {
     return activeStep === steps.length - 1
   }
